🚛 Rename: assets/header 폴더 생성 후 헤더 내  이미지 이동
